Aleksandrs K.: Lesson 14, rename package level_3 to level_3_4_5_6
